Adding encoder.yaml
It's not yet a particularly good example, but it is cut down from the running behemoth from ParallelEncoding, and it's pretty necessary. It's also now missing options that won't make any sense, and has a couple new ones. Not entirely sure it's actually, you know, the format I want yet, but may as well get something in.