Made Source Video optional, but enforced by yaml.
As it now stands, the harness can be used to run the lossless step from the yaml too, and it doesn't bother adding the needless argument to the global state of a video that doesn't yet exist.

Honestly, I'm probably going to have to support arbitrary -a passes at some point if I want this to be more generic, but it serves my needs now. It sends type to the script, which currently doesn't consume any python globals, but that's sort of okay. Also I suppose this also means it can be extended to more easily hardsub at the first stop.