Changed Exception Handling and removed regex
str.replace is significantly faster, and perfectly up to the only challenge it's asked to meet in this particular file. I honestly kind of forgot about it.

Also changed the exceptions that are supposed to kill things to kill more gently, with messages given to stderr instead of stdout, and reraising the initial exception instead of implicitly throwing a different one... though this may actually be worse since it's catching user error and throwing tracebacks, not program bugs. Time will tell with this one.