Change load_global_settings() to check __file__
I think ts_demux_audio.py is the only place that uses load_global_settings(), and it's also probably the only place that you're likely going to load one of the scripts as a library (because I don't know that you'd be able to sanely control the others), so it's important to ensure the library is looking in the right spot still.

This is probably bad practice overall, but it works?